EFB2RAM: Apply copy filter as a float coefficient after sampling

+++ Source/Core/VideoBackends/D3D/PSTextureEncoder.cpp
@@ -34,7 +34,7 @@ struct EFBEncodeParams
   float gamma_rcp;
   float clamp_top;
   float clamp_bottom;
-  s32 filter_coefficients[3];
+  float filter_coefficients[3];
   u32 padding;
 };
 
@@ -169,4 +169,4 @@ ID3D11PixelShader* PSTextureEncoder::GetEncodingPixelShader(const EFBCopyParams&
   m_encoding_shaders.emplace(params, newShader);
   return newShader;
 }
-}
+}  // namespace DX11
